Refactor: Flatten Scope and Disambiguation logic
- Flattened `Scope` in `src/semantic/resolver.rs` from a recursive `Box<Scope>` structure to a performant `Vec<ScopeLevel>` stack.
- Replaced `scope.child()` with `scope.enter()` and `scope.exit()` throughout the semantic analysis pipeline (`declarations.rs`, `control_flow.rs`), fixing variable leakage in loops.
- Flattened the module hierarchy by moving `DisambiguationContext` and scoring logic from `src/semantic/disambiguation.rs` to `src/morphology/mod.rs`.
- Moved `analyze_article` to `src/morphology/lexicon.rs` where it logically belongs.
- Deleted `src/semantic/disambiguation.rs` as it is no longer needed.
- Updated all call sites and verified with comprehensive tests.
