+++ src/semantic/mod.rs
@@ -35,17 +35,16 @@ pub mod assembler;
 pub(crate) mod control_flow;
 pub(crate) mod conversion;
 pub(crate) mod declarations;
-pub mod disambiguation;
 pub(crate) mod expressions;
 pub(crate) mod model;
 pub(crate) mod patterns;
 mod resolver;
 mod types;
 
+pub use crate::morphology::{DisambiguationContext, analyze_article, disambiguate, resolve_best};
 pub use assembler::{
     AssembledStatement, Assembler, AssemblyError, Constituent, Literal, VerbConstituent,
 };
-pub use disambiguation::{DisambiguationContext, analyze_article, disambiguate, resolve_best};
 pub use model::*;
 pub use resolver::*;
 pub use types::*;
